Update docker-compose and create Senrty, Dockerfile

+++ app/api_v1/products/routers.py
@@ -3,6 +3,8 @@ from typing import Any
 from fastapi import APIRouter, Depends, status
 from sqlalchemy.ext.asyncio import AsyncSession
 
+from app.api_v1.auth import AuthUser
+from app.api_v1.exceptions import HttpAPIException
 from app.api_v1.products.schemas import ProductShow, ProductCreate, ProductUpdate
 from app.api_v1.products.services import product_service
 from app.db import get_async_session
@@ -46,12 +48,16 @@ async def get_products(session: AsyncSession = Depends(get_async_session)) -> li
              status_code=status.HTTP_201_CREATED
              )
 async def create_product(product_data: ProductCreate,
-                         session: AsyncSession = Depends(get_async_session)) -> dict:
-    number_product: int = await product_service.add_product(session=session, product_data=product_data)
-    return {
-        "message": f"Product creates with {number_product} number successfully",
-        "data": product_data
-    }
+                         session: AsyncSession = Depends(get_async_session),
+                         user: dict = Depends(AuthUser.get_current_auth_user)) -> dict:
+    if user["is_superuser"] is True:
+        number_product: int = await product_service.add_product(session=session, product_data=product_data)
+        return {
+            "message": f"Product creates with {number_product} number successfully",
+            "data": product_data
+        }
+
+    raise HttpAPIException(exception="access denied.").http_error_403
 
 
 @router.delete(path="/{product_id}",
@@ -60,29 +66,26 @@ async def create_product(product_data: ProductCreate,
                )
 async def delete_product(
         product_id: int,
-        session: AsyncSession = Depends(get_async_session)) -> None:
-    await product_service.delete_product(id_product=product_id, session=session)
-    return None
+        session: AsyncSession = Depends(get_async_session),
+        user: dict = Depends(AuthUser.get_current_auth_user)) -> None:
+    if user["is_superuser"] is True:
+        await product_service.delete_product(id_product=product_id, session=session)
+        return None
+
+    raise HttpAPIException(exception="access denied.").http_error_403
 
 
-#
 @router.put(path="/{product_id}",
             summary='Обновление продукта'
             )
 async def update_product(
         product_id: int,
         product_update: ProductUpdate,
-        session: AsyncSession = Depends(get_async_session)) -> dict[str, Any]:
-    return await product_service.update_product(id_product=product_id,
-                                                session=session,
-                                                new_product=product_update)
-#
-#
-# @router.patch(path="/{id}",
-#               summary='Частичное обновление продукта',
-#               )
-# async def update_product_partial(
-#         product_update: ProductUpdatePartial,
-#         session: AsyncSession = Depends(get_async_session),
-#         product: ProductShow = Depends(product_by_id)) -> ProductShow:
-#     return await update_one_product_partial(session=session, product=product, product_update=product_update)
+        session: AsyncSession = Depends(get_async_session),
+        user: dict = Depends(AuthUser.get_current_auth_user)) -> dict[str, Any]:
+    if user["is_superuser"] is True:
+        return await product_service.update_product(id_product=product_id,
+                                                    session=session,
+                                                    new_product=product_update)
+
+    raise HttpAPIException(exception="access denied.").http_error_403
